requirements.txt
google-api-python-client
google-auth
pandas
openai
requests
weasyprint
beautifulsoup4
google-analytics-data
matplotlib
tabulate
playwright
lxml
reportlab
aiohttp
gspread
seaborn
numpy
